old-tests/submission/Microsoft/NavigationTimingL2/test_navigation_timing2_objects.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title></title>
<link rel="author" title="W3C" href="http://www.w3.org/" />
<link rel="help" href="http://www.w3.org/TR/navigation-timing-2/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css" />
<script>
test(function() {
assert_true(window.performance.getEntries()[0] instanceof PerformanceNavigationTiming);
}, "Verify type of navigation entry is PerformanceNavigationTiming using getEntries()");
test(function() {
assert_true(window.performance.getEntriesByName("document")[0] instanceof PerformanceNavigationTiming);
}, "Verify type of navigation entry is PerformanceNavigationTiming using getEntriesByName(\"document\")");
test(function() {
assert_true(window.performance.getEntriesByType("navigation")[0] instanceof PerformanceNavigationTiming);
}, "Verify type of navigation entry is PerformanceNavigationTiming using getEntriesByType(\"navigation\")");


</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that a Navigation Timing 2 entry can be accessed using the getEntries, getEntriesByType and getEntriesByName functions.</p>

<div id="log"></div>

</body>
</html>
